src/VoltageRegulator: Event Enabled/disabled do not reflect status
The 'event' signals a change in the state, but not the status.
Update the code to reflect that.
Fixes a power on failure when the CPU is too fast, the VR or I2C bus
too slow and status still reads as "off" after event enabled.

Also add a periodic timer that checks for the regulator status
mutliple times before signalling an error condition. This is required
as there's no IRQ when the regulator has powered on and voltage
is good. There might simply no sysfs notify and thus status must be
polled.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>