GpioInput: Add _Enabled signal
When the IO thread takes too long to execute the _On signal
is no good indicator to start the next IO. Wait for the IO
thread by introducing _Enabled signal, that is updated when
the IO thread is done.

Use the _Enabled signal to wait for IO thread before driving
SYS_PWROK. Fixes a timeout in PMC when it takes to long to
drive RESET_N due to long IO thread queue.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>