config/sbp1, GpioOutput: Wait for GPIO to be driven
Since GPIO driving has been moved to a separate thread it might take
too long for the GPIO to be seen by hardware. Add a new _Enabled signal
for each output GPIO to wait for the signal to be visible before continuing
sequencing.

Update the config to use the newly introduced signal.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>