config/sbp1: Add BCM0
Add a new signal <Name>_IsDummy to signal that a dummy regulator was
used. Allows to not active certain hardware groups and allows BCM0
to be missing.

Always enable BCM0 by default.
Enable BCM1-3 in chassis off when the previous regulator is a dummy.
In chassis on enable all regulators.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>